requirements.txt
argparse
numpy
Pillow
tensorflow
tensorflow_hub
